refactor(controller): bring integration tests and support inside the package
tests/controller/integration reached into the package from outside;
it now lives at controller/tests with its own relaxed-strictness
tsconfig for typed linting (the suite was previously typechecked by
nobody). The orphaned root fake-process-runner copy is removed, an
unused tempDir import fixed, engines field added, and docs updated —
controller/README documents the contracts package and drops the stale
prom-client claim.

Verified self-consistent: the controller directory copied outside the
repo installs, typechecks, passes 142 unit + 117 integration tests,
and boots /health=200 with zero references outside itself.
