+++ controller/README.md
@@ -17,7 +17,6 @@
 - Hono HTTP framework.
 - Effect Schema configuration and boundary validation.
 - SQLite-backed local stores.
-- `prom-client` metrics.
 - Swagger UI from `@hono/swagger-ui`.
 - Runtime probes for Python, Docker, `llama-server`, and MLX Python environments.
 
@@ -104,3 +103,4 @@ Runtime-related environment variables include:
 - `src/modules/proxy/`: OpenAI-compatible proxy and inference accounting.
 - `src/modules/system/`: metrics, logs, usage, events, and platform state.
 - `src/stores/`: SQLite helpers and persisted stores.
+- `contracts/`: the `@local-studio/contracts` package — the controller's HTTP API contract, consumed by the frontend via a `file:` dependency.
