+++ controller/eslint.config.mjs
@@ -59,6 +59,23 @@ const config = [
       ],
     },
   },
+  {
+    files: ["contracts/**/*.ts"],
+    rules: {
+      "unicorn/prevent-abbreviations": "off",
+    },
+  },
+  {
+    files: ["tests/**/*.ts"],
+    languageOptions: {
+      parserOptions: { project: "./tests/tsconfig.json" },
+    },
+    rules: {
+      "@typescript-eslint/explicit-function-return-type": "off",
+      "unicorn/prevent-abbreviations": "off",
+      "max-lines-per-function": "off",
+    },
+  },
 ];
 
 export default config;
