+++ controller/tests/integration/vllm-extra-args.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeEach, describe, expect, it } from "bun:test";
 
-import { appendExtraArguments } from "../../../controller/src/modules/engines/process/backend-builder";
-import { appendLlamacppArguments } from "../../../controller/src/modules/engines/specs/llamacpp-spec";
-import { appendVllmExtraArguments } from "../../../controller/src/modules/engines/specs/vllm-spec";
+import { appendExtraArguments } from "../../src/modules/engines/process/backend-builder";
+import { appendLlamacppArguments } from "../../src/modules/engines/specs/llamacpp-spec";
+import { appendVllmExtraArguments } from "../../src/modules/engines/specs/vllm-spec";
 
 describe("CLI extra_args serialization", () => {
   const extraArguments = {
